separated the logic into diffrent modules
